LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe-allow-expected.txt

CONSOLE MESSAGE: Permission policy 'Fullscreen' check failed for document with origin ''.
CONSOLE MESSAGE: Permission policy 'Fullscreen' check failed for document with origin ''.
CONSOLE MESSAGE: Permission policy 'Fullscreen' check failed for document with origin ''.
CONSOLE MESSAGE: Permission policy 'Fullscreen' check failed for document with origin ''.

PASS iframe-cross-origin-allow
